added contains() function and deleted Rectangle.java class
